+++ CHANGELOG.md
@@ -114,6 +114,7 @@ Modifications can only be done via builders.
 - Simplify overrides
 - Allow to submit connections when creating a scope from JSON
 - Rename customizer into builder
+- Rename findScope into findChildScope
 
 ### Fixed